+++ src/ru/yandex/practicum/delivery/model/StandardParcel.java
@@ -5,7 +5,7 @@ import ru.yandex.practicum.delivery.navigation.Address;
 import java.time.LocalDate;
 
 public class StandardParcel extends Parcel {
-    public StandardParcel(String description, double weight, Address deliveryAddress, LocalDate sendDay) {
+    public StandardParcel(String description, int weight, Address deliveryAddress, LocalDate sendDay) {
         super(description, weight, deliveryAddress, sendDay);
     }
 }
